requirements.txt
langchain>=0.2.11,<0.3
langchain-community>=0.2.11,<0.3
langchain-text-splitters>=0.2.2,<0.3
chromadb>=0.5.0,<0.6                        //版本老了
FlagEmbedding>=1.2.10,<2.0
pypdf>=4.0.2,<5.0
python-dotenv>=1.0.1,<2.0
typer>=0.12.3,<0.13
psutil>=5.9.8,<6.0
fastapi>=0.111.0,<0.112
uvicorn>=0.30.1,<0.31
click>=8.1.7,<9.0
streamlit>=1.37.0,<2.0
